requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
PyJWT==2.8.1
bcrypt==4.1.1
python-dotenv==1.0.0
pymongo==4.6.0
motor==3.3.2
email-validator==2.1.0
requests==2.31.0
boto3==1.29.7
onesignal-python-api==1.4.0
aiohttp==3.9.1
aiofiles==23.2.1
python-dateutil==2.8.2
pytz==2023.3
